Debug and release configuration of simple db and runtime use 64bit

+++ src/Amazon.Runtime/AmazonAbstractWebServiceClient.m
@@ -115,7 +115,7 @@
 
         response.isAsyncCall = NO;
         [self startSyncRequest:generatedRequest forRequest:urlRequest response:response originalRequest:originalRequest];
-        AMZLogDebug(@"Response Status Code : %d", response.httpStatusCode);
+        AMZLogDebug(@"Response Status Code : %ld", (long)response.httpStatusCode);
 
         if ( [self shouldRetry:response exception:((AmazonRequestDelegate *)generatedRequest.delegate).exception] ) {
             AMZLog(@"Retring Request: %d", retries);
